+++ CRM.DTOs/ProducOUCRDTOs/GetIdResultProductDTO.cs
@@ -17,6 +17,6 @@ namespace CRM.DTOs.ProducOUCRDTOs
         public string DescripcionOUCR { get; set; }
 
         [Display(Name = "Precio")]
-        public double PrecioOUCR { get; set; }
+        public decimal PrecioOUCR { get; set; }
     }
 }
